Add missing math.h intercept header
When building in BSL_OVERRIDES_STD mode, there should be an intercept
header for every standard header, but the <math.h> header was missing.

It is not clear that there is a real benefit intercepting the standard
C library headers, given that they should not interact with namespace
std/bsl, which is the whole reason to intercept.  However, adding the
missing file is the easiest way to consistency in the short term.

The fact that the lack of this patch did not fundamentally break any
code in Bloomberg suggests that the C intercept headers might be
safely removed.